feat: add haskey & has keys func

+++ pkg/client/backend.go
@@ -1,6 +1,76 @@
 package client
 
-import "context"
+import (
+	"context"
+	"errors"
+	"fmt"
+
+	"github.com/azrod/kivigo/pkg/errs"
+)
+
+// HasKey checks if the specified key exists in the store.
+// Returns an error if the key is empty.
+func (c Client) HasKey(ctx context.Context, key string) (bool, error) {
+	if key == "" {
+		return false, errs.ErrEmptyKey
+	}
+
+	_, err := c.GetRaw(ctx, key)
+	if err != nil {
+		if errors.Is(err, errs.ErrNotFound) {
+			return false, nil
+		}
+		return false, fmt.Errorf("failed to check key existence: %w", err)
+	}
+
+	return true, nil
+}
+
+// HasKeys checks if the specified keys exists in the store.
+// Returns an error if the keys is empty.
+func (c Client) HasKeys(ctx context.Context, keys []string) (bool, error) {
+	if len(keys) == 0 {
+		return false, errs.ErrEmptyKey
+	}
+
+	for _, key := range keys {
+		exists, err := c.HasKey(ctx, key)
+		if err != nil {
+			return false, fmt.Errorf("failed to check key existence: %w", err)
+		}
+		if !exists {
+			return false, nil
+		}
+	}
+
+	return true, nil
+}
+
+type MatchKeysFunc func(keys []string) (bool, error)
+
+// MatchKeys lists all keys with the given prefix and applies the provided MatchKeysFunc to determine if they match a custom condition.
+// Returns false and an error if the function is nil or listing keys fails.
+//
+// Example:
+//
+//	ok, err := client.MatchKeys(ctx, "prefix", func(keys []string) (bool, error) {
+//	    // Custom matching logic
+//	    return len(keys) > 0, nil
+//	})
+func (c Client) MatchKeys(ctx context.Context, prefix string, f MatchKeysFunc) (bool, error) {
+	if f == nil {
+		return false, errs.ErrEmptyFunc
+	}
+
+	// List all keys in the store
+	keys, err := c.List(ctx, prefix)
+	if err != nil {
+		return false, fmt.Errorf("failed to list keys: %w", err)
+	}
+
+	// Check if any key matches the criteria
+	return f(keys)
+}
 
 // Get retrieves the value stored under the specified key and decodes it into dest.
 // Returns an error if the key does not exist or decoding fails.
@@ -10,6 +80,10 @@ import "context"
 //	var value string
 //	err := client.Get(ctx, "myKey", &value)
 func (c Client) Get(ctx context.Context, key string, value any) error {
+	if key == "" {
+		return errs.ErrEmptyKey
+	}
+
 	vV, err := c.GetRaw(ctx, key)
 	if err != nil {
 		return err
@@ -25,6 +99,10 @@ func (c Client) Get(ctx context.Context, key string, value any) error {
 //
 //	err := client.Set(ctx, "myKey", "myValue")
 func (c Client) Set(ctx context.Context, key string, value any) error {
+	if key == "" {
+		return errs.ErrEmptyKey
+	}
+
 	vV, err := c.opts.Encoder.Encode(value)
 	if err != nil {
 		return err
@@ -50,6 +128,10 @@ func (c Client) Set(ctx context.Context, key string, value any) error {
 //
 //	err := client.Delete(ctx, "myKey")
 func (c Client) Delete(ctx context.Context, key string) error {
+	if key == "" {
+		return errs.ErrEmptyKey
+	}
+
 	err := c.KV.Delete(ctx, key)
 	if err != nil {
 		return err
